[FIX] : 파이어베이스 토큰 부분 수정 메인페이지로 옮김
