+++ app/src/main/java/com/teambeme/beme/BeMeApplication.kt
@@ -12,19 +12,6 @@ class BeMeApplication : Application() {
         super.onCreate()
         initPixelUtil()
         BeMeAuthPreference.init(this)
-        FirebaseMessaging.getInstance().token.addOnCompleteListener(OnCompleteListener { task ->
-            if (!task.isSuccessful) {
-                Log.w("BeMeApplication.TAG", "Fetching FCM registration token failed", task.exception)
-                return@OnCompleteListener
-            } else {
-                // Get new FCM registration token
-                val token = task.result
-                BeMeAuthPreference.fireBaseToken = token ?: "SomeThing"
-                // Log and toast
-                val msg = getString(R.string.msg_token_fmt, token)
-                Log.d("BeMeApplication.TAG", msg)
-            }
-        })
     }
 
     private fun initPixelUtil() {
