+++ app/src/main/java/com/teambeme/beme/idsearchfollowing/adapter/IdSearchAdapter.kt
@@ -119,8 +119,5 @@ class IdSearchAdapter(
             binding.btnFollowingFollow.visibility = View.VISIBLE
         }
     }
-
-
-
 }
